Modify: 재료 리스트, 재료 리스트 아이템 컴포넌트를 하나로 합침
