+++ src/components/Mypage/IngredientList.tsx
@@ -1,22 +1,32 @@
 import { css } from "@styled-system/css";
 
-import IngredientListItem from "./IngredientListItem";
+import { IngredientKey, IngredientQuantity } from "@/types/ingredient";
 
-const IngredientList = () => {
+import { INGREDIENT_ICON_BY_KEY, INGREDIENT_NAME_BY_KEY } from "@/constants/ingredient";
+
+interface Props {
+  ingredients: IngredientQuantity[];
+  uniqueIngredient: IngredientKey;
+}
+
+const INFINITY = "\u221E";
+
+const IngredientList = ({ ingredients, uniqueIngredient }: Props) => {
   return (
     <ul className={styles.list}>
-      <IngredientListItem ingredientCount={1} />
-      <IngredientListItem ingredientCount={2} />
-      <IngredientListItem ingredientCount={3} />
-      <IngredientListItem ingredientCount={4} />
-      <IngredientListItem ingredientCount={5} />
-      <IngredientListItem ingredientCount={6} />
-      <IngredientListItem ingredientCount={7} />
-      <IngredientListItem ingredientCount={8} />
-      <IngredientListItem ingredientCount={9} />
-      <IngredientListItem ingredientCount={10} />
-      <IngredientListItem ingredientCount={11} />
-      <IngredientListItem ingredientCount={12} />
+      {ingredients.map(({ ingredient, stockQuantity }, index) => {
+        const Icon = INGREDIENT_ICON_BY_KEY[ingredient];
+
+        return (
+          <li key={`${index}-${ingredient}`} className={styles.item}>
+            <div className={styles.title}>{INGREDIENT_NAME_BY_KEY[ingredient]}</div>
+            <div className={styles.iconContainer}>
+              <Icon />
+              <span>X {ingredient === uniqueIngredient ? INFINITY : stockQuantity}</span>
+            </div>
+          </li>
+        );
+      })}
     </ul>
   );
 };
@@ -34,5 +44,25 @@ const styles = {
     overflowX: "auto",
     marginTop: "1.2rem",
     marginBottom: "3.2rem",
+  }),
+  item: css({
+    width: "9.6rem",
+    height: "9.3rem",
+    backgroundColor: "primary.20",
+    borderRadius: "0.8rem",
+  }),
+  title: css({
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    height: "2.9rem",
+    backgroundColor: "primary.100",
+    borderTopRadius: "0.8rem",
+  }),
+  iconContainer: css({
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    height: "6.4rem",
   }),
 };
